# [libc++] Optimize ranges::fill{,_n} for vector<bool>::iterator